refactor: remove FiltersBar de dentro da List
