--- src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/FiltersBar/modules/common/SelectConfig.ts
@@ -1,12 +1,0 @@
-export type HtmlSelectOption = {
-    value: string;
-    label: string;
-};
-
-export type SelectConfig = {
-    id: string;
-    label: string;
-    options: Array<HtmlSelectOption>;
-    isMultiSelect: boolean;
-    width: string;
-};
