+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/common/toSelectConfigsShared/index.ts
@@ -1,4 +1,3 @@
-import { referenceOrder } from "./consts";
 import type * as z from "zod/v4";
 import type {
     ConditionIdentifiedBy,
@@ -10,7 +9,17 @@ import { nameFormatter } from "@/features/acf/frontend/diabetes/modules/AcfPage/
 import type {
     HtmlSelectOption,
     SelectConfig,
-} from "./modules/common/SelectConfig";
+} from "@/features/acf/frontend/common/SelectConfig";
+
+const referenceOrder = [
+    "Menos de 17 anos",
+    "Entre 18 e 24 anos",
+    "Entre 25 e 34 anos",
+    "Entre 35 e 44 anos",
+    "Entre 45 e 54 anos",
+    "Entre 55 e 65 anos",
+    "Mais de 65 anos",
+];
 
 export type FilterOptions =
     | Array<string>
