+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/print/buildPrintProps.ts
@@ -9,7 +9,7 @@ import {
 import type { FilterItem } from "@/services/lista-nominal/ListaNominal";
 import { PROFILE_ID, type ProfileIdValue } from "@/types/profile";
 import type { DiabetesAcfPrintItem } from "./diabetes/print.model";
-import { filtersLabels } from "../FiltersBar/modules/CoeqFiltersBar/consts";
+import { filtersLabels } from "../CoeqFiltersBar/consts";
 import { coeqColumns } from "../CoeqDataTable/consts";
 
 export const buildPrintProps = (
